Improved the transitions
If a new component is added to a keyframe but with no start side set, it
will be automatically added to the panel at its target location after
the transition. Similarly, if an existing component is not part of a
keyframe (ie. it is referred as an old component) and have no end side
set, it will be immedialety removed from the panel at the beginning of
the transition.

Also, components that slide out of the panel are removed after the
transition. It should improve the performances.